examples/mcp_vibe_example/pyproject.toml
[project]
name = "mcp-vibe-example"
version = "0.1.0"
description = "Self-contained MCP server example using gRPC VibeService"
requires-python = ">=3.10"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Stable Kernel", email = "info@stablekernel.com"}
]
keywords = ["mcp", "grpc", "protobuf", "example", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Communications",
]

dependencies = [
    "fastmcp>=2.3.1",
    "protobuf>=5.26.1,<6.0",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
]

[project.scripts]
mcp-vibe = "mcp_vibe_app.server:main"

[project.urls]
Repository = "https://github.com/stablekernel/protoc-gen-py-mcp"
Documentation = "https://github.com/stablekernel/protoc-gen-py-mcp/tree/main/examples/mcp_vibe_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_vibe_app", "protos"]

[tool.hatch.build.targets.sdist]
include = [
    "/mcp_vibe_app",
    "/protos",
    "/README.md",
    "/pyproject.toml",
]